# samples/maven-producer/src/main/java/org/example/order/OrderController.java
/*
 * Copyright 2026-present the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.example.order;

import java.util.Map;
import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.atomic.AtomicLong;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/orders")
public class OrderController {

	private final Map<String, Order> orders = new ConcurrentHashMap<>();

	private final AtomicLong idGenerator = new AtomicLong(1);

	public OrderController() {
		this.orders.put("1", new Order("1", "MacBook Pro", 1299.99, "CREATED"));
	}

	@GetMapping("/{id}")
	public ResponseEntity<Order> getOrder(@PathVariable String id) {
		Order order = this.orders.get(id);
		if (order == null) {
			return ResponseEntity.notFound().build();
		}
		return ResponseEntity.ok(order);
	}

	@PostMapping
	public ResponseEntity<Order> createOrder(@RequestBody CreateOrderRequest request) {
		String id = String.valueOf(this.idGenerator.incrementAndGet());
		Order order = new Order(id, request.product(), request.amount(), "CREATED");
		this.orders.put(id, order);
		return ResponseEntity.status(201).body(order);
	}

	public record CreateOrderRequest(String product, double amount) {
	}

}